// apps/web/src/lib/requests/api-client/model/paymentResponseDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * API Docs
 * API description
 * OpenAPI spec version: 1.0
 */
import type { PaymentStatus } from './paymentStatus';
import type { PaymentResponseDtoDepositDetails } from './paymentResponseDtoDepositDetails';

export interface PaymentResponseDto {
  id: string;
  amount: string;
  currency: string;
  status: PaymentStatus;
  createdAt: string;
  expiredAt: string;
  /** @nullable */
  depositDetails: PaymentResponseDtoDepositDetails;
  paymentUrl: string;
}